chore: updated properties script to preserve existing configurations

+++ properties.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
-
+echo
 ls | grep ecourse. | while read line; do
   path="$line/src/main/resources"
   if [ -f "$path/application.example.properties" ]; then
-    cp "$path/application.example.properties" "$path/application.properties"
+    if [ -f "$path/application.properties" ]; then
+      echo "> $path already has application.properties. Skipping..."
+    else
+      echo "> Creating $path/application.properties"
+      cp "$path/application.example.properties" "$path/application.properties"
+    fi
   fi
 done
